perf(ext/node): convert 15 more polyfill files to lazy-loaded scripts
Converts 9 node:* entry point modules and 3 path internals from ESM to
IIFE-wrapped lazy-loaded scripts, with thin ESM shims providing the
module interface.

Converted node:* modules (new script + ESM shim):
- async_hooks, trace_events, wasi, perf_hooks, sqlite
- querystring, string_decoder, vm, diagnostics_channel

Converted path internals (in-place to scripts):
- path/_posix.ts, path/_win32.ts, path/mod.ts

Updated path entry point shims:
- path.ts, path/posix.ts, path/win32.ts
