+++ ext/node/lib.rs
@@ -485,9 +485,6 @@ deno_core::extension!(deno_node,
     "internal/util/comparisons.ts",
     "internal/util/parse_args/parse_args.js",
     "internal/webstreams/adapters.js",
-    "path/_posix.ts",
-    "path/_win32.ts",
-    "path/mod.ts",
     "readline/promises.ts",
     "_fs/_fs_common.ts",
     "_fs/_fs_cp.ts",
@@ -578,6 +575,15 @@ deno_core::extension!(deno_node,
   ],
   lazy_loaded_js = [
     dir "polyfills",
+    "_async_hooks.ts",
+    "_diagnostics_channel.js",
+    "_perf_hooks.js",
+    "_querystring.js",
+    "_sqlite.ts",
+    "_string_decoder.ts",
+    "_trace_events.ts",
+    "_vm.js",
+    "_wasi.ts",
     "internal/validators.mjs",
     "internal/normalize_encoding.ts",
     "internal/error_codes.ts",
@@ -654,8 +660,11 @@ deno_core::extension!(deno_node,
     "internal/util/parse_args/utils.js",
     "path/_constants.ts",
     "path/_interface.ts",
+    "path/_posix.ts",
     "path/_util.ts",
+    "path/_win32.ts",
     "path/common.ts",
+    "path/mod.ts",
     "path/separator.ts",
   ],
   options = {
