+++ ext/node/polyfills/path/_win32.ts
@@ -2,11 +2,8 @@
 // Ported from https://github.com/browserify/path-browserify/
 // Copyright 2018-2026 the Deno authors. MIT license.
 
-import type {
-  FormatInputPathObject,
-  ParsedPath,
-} from "ext:deno_node/path/_interface.ts";
-import { core, primordials } from "ext:core/mod.js";
+(function () {
+const { core, primordials } = globalThis.__bootstrap;
 const {
   CHAR_BACKWARD_SLASH,
   CHAR_COLON,
@@ -25,11 +22,37 @@ const {
   isWindowsDeviceRoot,
   normalizeString,
 } = core.loadExtScript("ext:deno_node/path/_util.ts");
-import assert from "node:assert";
-import process from "node:process";
-import type * as fsGlob from "ext:deno_node/_fs/_fs_glob.ts";
+const { default: assert } = core.loadExtScript(
+  "ext:deno_node/internal/assert.mjs",
+);
+const _proc = core.loadExtScript("ext:deno_node/_process/process.ts");
+const process = { cwd: _proc.cwd, env: _proc.env };
 
-const lazyLoadGlob = core.createLazyLoader<typeof fsGlob>(
+type FormatInputPathObject = {
+  root?: string;
+  dir?: string;
+  base?: string;
+  ext?: string;
+  name?: string;
+};
+
+type ParsedPath = {
+  root: string;
+  dir: string;
+  base: string;
+  ext: string;
+  name: string;
+};
+
+type FsGlob = {
+  matchGlobPattern: (
+    path: string,
+    pattern: string,
+    isWindows: boolean,
+  ) => boolean;
+};
+
+const lazyLoadGlob = core.createLazyLoader<FsGlob>(
   "ext:deno_node/_fs/_fs_glob.ts",
 );
 
@@ -49,8 +72,8 @@ const {
   TypeError,
 } = primordials;
 
-export const sep = "\\";
-export const delimiter = ";";
+const sep = "\\";
+const delimiter = ";";
 
 const WINDOWS_RESERVED_NAMES = [
   "CON",
@@ -94,7 +117,7 @@ function isWindowsReservedName(path: string, colonIndex: number): boolean {
  * Resolves path segments into a `path`
  * @param pathSegments to process to path
  */
-export function resolve(...pathSegments: string[]): string {
+function resolve(...pathSegments: string[]): string {
   let resolvedDevice = "";
   let resolvedTail = "";
   let resolvedAbsolute = false;
@@ -278,7 +301,7 @@ export function resolve(...pathSegments: string[]): string {
  * Normalizes a `path`
  * @param path to normalize
  */
-export function normalize(path: string): string {
+function normalize(path: string): string {
   assertPath(path);
   const len = path.length;
   if (len === 0) return ".";
@@ -441,7 +464,7 @@ export function normalize(path: string): string {
  * Verifies whether path is absolute
  * @param path to verify
  */
-export function isAbsolute(path: string): boolean {
+function isAbsolute(path: string): boolean {
   assertPath(path);
   const len = path.length;
   if (len === 0) return false;
@@ -463,7 +486,7 @@ export function isAbsolute(path: string): boolean {
  * Join all given a sequence of `paths`,then normalizes the resulting path.
  * @param paths to be joined and normalized
  */
-export function join(...paths: string[]): string {
+function join(...paths: string[]): string {
   const pathsCount = paths.length;
   if (pathsCount === 0) return ".";
 
@@ -568,7 +591,7 @@ export function join(...paths: string[]): string {
  * @param from relative path
  * @param to relative path
  */
-export function relative(from: string, to: string): string {
+function relative(from: string, to: string): string {
   assertPath(from);
   assertPath(to);
 
@@ -743,7 +766,7 @@ export function relative(from: string, to: string): string {
  * Resolves path to a namespace path
  * @param path to resolve to namespace
  */
-export function toNamespacedPath(path: string): string {
+function toNamespacedPath(path: string): string {
   // Note: this will *probably* throw somewhere.
   if (typeof path !== "string") return path;
   if (path.length === 0) return "";
@@ -780,7 +803,7 @@ export function toNamespacedPath(path: string): string {
  * Return the directory name of a `path`.
  * @param path to determine name for
  */
-export function dirname(path: string): string {
+function dirname(path: string): string {
   assertPath(path);
   const len = path.length;
   if (len === 0) return ".";
@@ -875,7 +898,7 @@ export function dirname(path: string): string {
  * @param path to process
  * @param ext of path directory
  */
-export function basename(path: string, ext = ""): string {
+function basename(path: string, ext = ""): string {
   if (ext !== undefined && typeof ext !== "string") {
     throw new ERR_INVALID_ARG_TYPE("ext", ["string"], ext);
   }
@@ -964,7 +987,7 @@ export function basename(path: string, ext = ""): string {
  * Return the extension of the `path`.
  * @param path with extension
  */
-export function extname(path: string): string {
+function extname(path: string): string {
   assertPath(path);
   let start = 0;
   let startDot = -1;
@@ -1032,7 +1055,7 @@ export function extname(path: string): string {
  * Generate a path from `FormatInputPathObject` object.
  * @param pathObject with path
  */
-export function format(pathObject: FormatInputPathObject): string {
+function format(pathObject: FormatInputPathObject): string {
   if (pathObject === null || typeof pathObject !== "object") {
     throw new ERR_INVALID_ARG_TYPE("pathObject", ["Object"], pathObject);
   }
@@ -1043,7 +1066,7 @@ export function format(pathObject: FormatInputPathObject): string {
  * Return a `ParsedPath` object of the `path`.
  * @param path to process
  */
-export function parse(path: string): ParsedPath {
+function parse(path: string): ParsedPath {
   assertPath(path);
 
   const ret: ParsedPath = { root: "", dir: "", base: "", ext: "", name: "" };
@@ -1192,15 +1215,17 @@ export function parse(path: string): ParsedPath {
   return ret;
 }
 
-export const _makeLong = toNamespacedPath;
+const _makeLong = toNamespacedPath;
 
-let lazyMatchGlobPattern: typeof fsGlob.matchGlobPattern;
-export const matchesGlob = (path: string, pattern: string): boolean => {
+let lazyMatchGlobPattern:
+  | ((path: string, pattern: string, isWindows: boolean) => boolean)
+  | undefined;
+const matchesGlob = (path: string, pattern: string): boolean => {
   lazyMatchGlobPattern ??= lazyLoadGlob().matchGlobPattern;
   return lazyMatchGlobPattern(path, pattern, true);
 };
 
-export default {
+const defaultExport = {
   basename,
   delimiter,
   dirname,
@@ -1217,3 +1242,24 @@ export default {
   _makeLong,
   matchesGlob,
 };
+
+return {
+  basename,
+  delimiter,
+  dirname,
+  extname,
+  format,
+  isAbsolute,
+  join,
+  normalize,
+  parse,
+  relative,
+  resolve,
+  sep,
+  toNamespacedPath,
+  _makeLong,
+  matchesGlob,
+  default: defaultExport,
+  "module.exports": defaultExport,
+};
+})();
