+++ ext/node/polyfills/wasi.ts
@@ -1,15 +1,6 @@
 // Copyright 2018-2026 the Deno authors. MIT license.
 
-import { primordials } from "ext:core/mod.js";
-
-const { Error } = primordials;
-
-class Context {
-  constructor() {
-    throw new Error("Context is currently not supported");
-  }
-}
-
-export const WASI = Context;
-
-export default { WASI };
+import { core } from "ext:core/mod.js";
+const _mod = core.loadExtScript("ext:deno_node/_wasi.ts");
+export const { WASI } = _mod;
+export default _mod.default;
